Add files via upload
The goal of this project was to create a website that would randomly select a movie based on the genre the user selects. A catalogue of all movies from the 1900s to 2018 was imported from IMBD and transferred into a MSSQL database. When the user was given a movie, a poster would be generated using a movie poster API, which was my main involvement in this project.